requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
anthropic>=0.12.0
